Refactoring ort fusions folder
