+++ onnxscript/rewriter/ort_fusions/_core.py
@@ -4,12 +4,12 @@ from __future__ import annotations
 
 import onnxscript.ir as ir
 from onnxscript.optimizer import remove_unused_nodes, optimize
-from onnxscript.rewriter.onnxruntime.xformers.cos_sin_cache import fuse_cos_sin_cache
-from onnxscript.rewriter.onnxruntime.xformers.mha import fuse_mha
-from onnxscript.rewriter.onnxruntime.xformers.rms_normalization import fuse_rms_normalization
-from onnxscript.rewriter.onnxruntime.xformers.rotary_embedding import fuse_rotary_embedding
-from onnxscript.rewriter.onnxruntime.xformers.sdpa import fuse_sdpa
-from onnxscript.rewriter.onnxruntime.xformers.skip_normalization import fuse_normalization
+from onnxscript.rewriter.ort_fusions.cos_sin_cache import fuse_cos_sin_cache
+from onnxscript.rewriter.ort_fusions.mha import fuse_mha
+from onnxscript.rewriter.ort_fusions.rms_normalization import fuse_rms_normalization
+from onnxscript.rewriter.ort_fusions.rotary_embedding import fuse_rotary_embedding
+from onnxscript.rewriter.ort_fusions.sdpa import fuse_sdpa
+from onnxscript.rewriter.ort_fusions.skip_normalization import fuse_normalization
 
 
 def fuse_xformers(model: ir.Model) -> None:
