+++ onnxscript/rewriter/ort_fusions/_smollm_1.py
