+++ onnxscript/rewriter/ort_fusions/_smollm_2.py
