+++ onnxscript/rewriter/ort_fusions/rms_normalization_test.py
@@ -5,22 +5,21 @@ from __future__ import annotations
 import unittest
 
 import onnxscript.optimizer
-from onnxscript.rewriter.onnxruntime.xformers import fuse_cos_sin_cache, fuse_rotary_embedding
-from onnxscript.rewriter.onnxruntime.xformers._smollm_1 import TestData
-from onnxscript.rewriter.onnxruntime.xformers._test_utils import assert_allclose, ort_run
+from onnxscript.rewriter.ort_fusions._smollm_1 import TestData
+from onnxscript.rewriter.ort_fusions._test_utils import assert_allclose, ort_run
+from onnxscript.rewriter.ort_fusions.rms_normalization import fuse_rms_normalization
 
 
-class TestCosSinCacheTransform(unittest.TestCase):
+class TestRmsNormalization(unittest.TestCase):
     def test_smollm(self):
         smollm_test = TestData()
         model = smollm_test.get_onnx_model()
         onnxscript.optimizer.optimize(model)
         inputs = smollm_test.get_ort_inputs()
         original_outputs = ort_run("original", model, inputs)
-        count = fuse_rotary_embedding(model)
-        self.assertGreater(count, 0)
-        count = fuse_cos_sin_cache(model)
-        self.assertGreater(count, 0)
+        fuse_rms_normalization(model)
+        op_types = [n.op_type for n in model.graph]
+        self.assertIn("SimplifiedLayerNormalization", op_types)
         new_outputs = ort_run("optimized", model, inputs)
         assert_allclose(new_outputs, original_outputs)
